Add first cut of a script that retrieves topology from ESDB. The topology data is then postprocessed and put into the files that OSCARS expects to read to get the network topology.
Based on esnet_topo.py in this directory.

Running this script requires an API token for ESDB.  It is
specified in one of four ways, in this order:

1.  Literally via the --token argument

2.  Literally in the ESDB_TOKEN environment variable.

3.  In the file specified by the --token-path argument.

4.  In the file specified by the ESDB_TOKEN_PATH environement
variable.

This script also uses some of the Python scripts (as a library)
contained in esnet/nesg-tools.  The Python library search path
($PYTHONPATH) needs to point to a copy of esnet/nesg-tools/esnet-python.

Towards #84.