Towards #91 (Make OSCARS poll for new / updated topology files).
Periodically (every hour) do a call to ESDB to grab the current "today" topology.  This requires a valid ESDB API key in the topo.esdb-key property.

Note that all this change does is update files.  It doesn't force OSCARS to process the updated topology, or make changes to its internal topology model.  That functionality is slated for issue #92.